Break meta tests into multiple files
