+++ tests/meta/test_ConceptWrapper.cc
@@ -1,0 +1,9 @@
+#define CATCH_CONFIG_MAIN
+#include "third-party/Catch/single_include/catch.hpp"
+
+#include "meta/ConceptWrapper.h"
+
+TEST_CASE("Test ConceptWrapper", "[games]")
+{
+
+}
