+++ tests/meta/test_StringType.cc
@@ -1,0 +1,9 @@
+#define CATCH_CONFIG_MAIN
+#include "third-party/Catch/single_include/catch.hpp"
+
+#include "meta/StringType.h"
+
+TEST_CASE("Test StringType", "[games]")
+{
+
+}
