+++ tests/meta/test_ValPack.cc
@@ -1,0 +1,9 @@
+#define CATCH_CONFIG_MAIN
+#include "third-party/Catch/single_include/catch.hpp"
+
+#include "meta/ValPack.h"
+
+TEST_CASE("Test ValPack", "[games]")
+{
+
+}
