+++ tests/meta/test_type_traits.cc
@@ -1,0 +1,9 @@
+#define CATCH_CONFIG_MAIN
+#include "third-party/Catch/single_include/catch.hpp"
+
+#include "meta/type_traits.h"
+
+TEST_CASE("Test type_traits", "[games]")
+{
+
+}
